requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.1.0
python-dotenv>=1.0.0
requests>=2.28.2
pydantic[email]>=2.0.0
pyyaml>=6.0
python-slugify>=8.0.1
cryptography>=40.0.0
opensearch-py>=2.0.0
openapy>=0.0.1
alembic>=1.11.0
